soc/google/cpif: Fix power_stats total sleep usec decrementing on modem reset
When the modem crashes or undergoes a silent reset, the
cp2ap_suspend_work gets called twice and on the next invocation of
cp2ap_wakeup_work after modem comes online, the suspend duration
gets a lower value than before the crash. This temporary backward
jump in value causes the PowerStats service to create an alarm.
We are adding a suspend flag to fix the issue so that the when the
suspend function is invoked twice, the entry timestamp is not
updated from last value.

Bug: 242524598
Signed-off-by: Mahesh Kallelil <kallelil@google.com>
Change-Id: I61415cf6b943b5b7aa913443085bc9cf0633a7fe
(cherry picked from commit fc8bda9e839463456ca104e763e11420995c5fc1)