exec: simplify initial stack size expansion
I had a hard time trying to understand completely why it is using vm_end in
one side of the expression and vm_start in the other one, and using
something in the "if" clause that is not an exact copy of what is used
below. The whole point is that the stack_size variable that was used in the
"if" clause is the difference between vm_start and vm_end, which is not far
away but makes this thing harder to read than it must be.

Signed-off-by: Rolf Eike Beer <eb@emlix.com>
Signed-off-by: Kees Cook <keescook@chromium.org>
Link: https://lore.kernel.org/r/2017429.gqNitNVd0C@mobilepool36.emlix.com
(cherry picked from commit bfb4a2b95875a47a01234f2de113ec089d524e71)
(cherry picked from commit fa30f89a9a3b72e78cf8acb67d498685e6d90e19)
(cherry picked from commit ce70eac55dc16e507e2e5c971738469c9f6fd45c)