vh: sched: add new vendor hook update_misfit_status
If a misfit task is affined to a subset of the possible cpus, we need
to verify that one of these cpus can fit it. Otherwise the load balance
code will continuously trigger needlessly leading to balance_interval to
increase in return and eventually end up with a situation where real
imbalances take a long time to address because of this impossible
imbalance situation.

Bug: 261701366
Change-Id: Ifffa2e239a0c960b21b4738a76bb7b8e9343d1f6
Signed-off-by: Chungkai Mei <chungkai@google.com>