vh: sched: Implement full performance inheritance in rt_mutex
Inheriting uclamp_min only makes sure we inherit the boost value. But
the utilization signal of the task is the true representation of the
performance requirement.

For example if

	p_uclamp_min = 0
	pi_uclamp_min = 0

but

	p_util = 100
	pi_util = 800

then the pi_task will have to wait for longer for p to release the lock
because of performance inversion. Inheriting uclamp_min value will not
help in these cases.

BUG: 242673160
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I071bfb37e16e0154aff20bdc27c33e4a48f49cbd
(cherry picked from commit 655075086e88e70b1da862d01bab3032f3ed6ee5)