vh: sched: Update uclamp_max/min filter at tick
Taken almost as-is from aosp/1914696.

One difference is that we need to ensure we clear UCLAMP_IDLE flag at
tick. Idle handling in the original proposal was different and that was
not ported.

We use update_misfit_status() vendor hook instead of scheduler tick as
it is called from tick_task_fair() and more equivalent to what is in
aosp/1914696.

Bug: 261695814
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Ide3c8d2a31aa508bd29cc789c272afcd2d008795
(cherry picked from commit 8072ec7caf1804244c03ec12726f2b1e8a0e4b67)