vh: sched: Block cpufreq updates until uclamp filters are applied
We teach sched_pixel how to ignore cpufreq update requests for a short
period of time.

This is required to ensure request to update frequency made by
enqueue_task_fair() which can happen before the filter is applied don't
result in applying wrong frequencies. Which not only could lead to
unnecessary spike in frequency, but our request to update the frequency
later after applying the filter could be ignored because of the rate
limit.

This addresses the issue of iowait boost causing unnecessary frequency
request which could lead to our request to be dropped because of the
rate limit too.

It should be unlikely to have multiple caller of block/resume. But we
use a counter just in case we end up nesting somehow.

Bug: 261695814
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I59d574d2bebe215c55a2dc018dbe2f80f222783f
(cherry picked from commit 44070558c0071cff7342580f4528ec0e7fc201ad)