vh: sched: Implement auto_uclamp_max
uclamp_max is an absolute value based on total system performance. But
on HMP systems with multiple clusters, the most efficient value is hard
to describe based on the cluster.

For example, if 512 is the most efficient point on the mid cluster, it
can still lead the task to run at max frequency on the little cluster as
its capacity value is usually a lot lower than that.

What we want instead for uclamp_max to be set automatically for the most
efficient point for the cluster the task is running on. And for this to
shift dynamically as the scheduler changes the placement of the task.

This helps in two ways:

	1. Userspace won't have to do guess work to find the most energy
	   efficient point for any particular system (portability).
	2. We won't accidentally use in-efficient OPPs.

It has a drawback though:

	1. Task placement hint is lost. uclamp_max is only a frequency
	   hint. If one want to bias task placement to stay on
	   a specific cluster, one shouldn't use the auto_uclamp_max.

Limitations:

	1. Only per task and vendor group implementation. No cgroup
	   support.
	2. We always use task_cpu() when looking up the effective
	   uclamp_max when auto_uclamp_max is set.
	   Only exception is in feec() where we consider uclamp_max
	   based on the cpu we are evaluating. Load balancer and
	   select_idle_capacity() will see effective uclamp_max based on
	   its current placement only.

Bug: 283761535
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I9b002e201e2fb93e932bfd5edef4d1be71e561f8
(cherry picked from commit b9d23761a6238a970c5cb3fd6ebdde9fceb0bd34)
(cherry picked from commit 54752a1949f446c5a7616665b0630ffebaa56b24)