vh: sched: Don't block freq updates at enqueue
For some reason we still see regressions on use cases that relies on
iowait boost.

Tried several things but couldn't get the lost perf back. So as last
resort reworking the logic so that we don't block the updates, but
rather always force an additional freq update after applying the uclamp
filters. Not ideal, but should retain old behavior, and ensure necessary
freq update are applied to impose uclamp hints - albeit we might have
already moved to wrong frequency temporarily.

Bug: 281263677
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Ib288c014ac652c1d482afda7ff60a2cae5600102