Revert "ANDROID: Update the ABI representation"
This reverts commit 4299d5b3e6b381c92b2176cb8c8602852f594278.

Reason for revert:

While syncing android13-gs-pixel-5.10-udc to GKI monthly branch android13-5.10-2023-03, these changes in the relation chain are not exist on android13-5.10-2023-03
(which means the sync point already exceed the initial SHA of android13-5.10-2023-03).
Therefore, revert the original changes on udc (and udc-qpr1) at that time.
For now syncing android13-gs-pixel-5.10-udc-qpr1 to android13-5.10-2023-07 and these reverted changes are needed to put back before the aosp merge.

This change is to keep the consistency of android/abi_gki_aarch64.xml with the following changes in the relation chain and avoid the conflict in the following merge process.

Bug: 281607159
Bug: 285104297
Bug: 283761535
Change-Id: Ib266abd037bbc455369facc4d012afb92f3f414a
Signed-off-by: Robin Peng <robinpeng@google.com>